Feature/stabilize api for february

+++ Fanview.API/Repository/Interface/ITeamLiveStatusRepository.cs
@@ -1,15 +1,13 @@
 ﻿using Fanview.API.Model;
 using MongoDB.Driver;
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace Fanview.API.Repository.Interface
 {
     public interface ITeamLiveStatusRepository
     {
-        Task<int> GetTeamLiveStatusCount(string matchId);
+        Task<long> GetTeamLiveStatusCount(string matchId);
 
         void CreateTeamLiveStatus(IList<LiveMatchStatus> teamLiveStatuses);
